refactor(desktop): split the settings panel into its pages
3,425 lines in one file, holding the controls' own types, every shared row, all
four pages, and the panel that chooses between them. Each is now the file it
was always a section of, and `pageResetControl`'s three identical branches are
one table over the scopes.
