+++ apps/desktop/src/renderer/settings/notes.ts
@@ -1,0 +1,7 @@
+/* Why a row that could otherwise be connected is not offering to be. */
+export const HELD_TITLE = "Finish the one you are entering first";
+
+/* Why every Connect in a key-holding section is refusing, said once per
+   section: a disabled control with no words beside it reads as broken. */
+export const STORAGE_UNAVAILABLE_NOTE =
+  "This system offers no encrypted credential storage, so Luke will not store a key here.";
